ProtoToBQ/requirements.txt
protobuf==3.13.0
six==1.15.0
google-cloud-pubsub>=0.39.0